weather_api/requirements.txt
click==8.1.3
Flask==2.2.2
flask_sqlalchemy==3.0.2
flask_migrate==4.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
Werkzeug==2.2.2
python-dotenv==0.21.0
